requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiofiles==24.1.0
    # via aiogram
aiogram==3.13.1
    # via -r requirements.in
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via aiogram
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.20.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
attrs==24.2.0
    # via aiohttp
certifi==2024.8.30
    # via aiogram
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
magic-filter==1.0.12
    # via aiogram
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
propcache==0.2.0
    # via yarl
pydantic==2.9.2
    # via aiogram
pydantic-core==2.23.4
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   aiogram
    #   aiosqlite
    #   pydantic
    #   pydantic-core
yarl==1.17.0
    # via aiohttp
